feat: 重新优化内置MJ结构
