+++ nodemon.json
@@ -1,5 +1,5 @@
 {
-  "watch": ["server/*.ts"],
-  "exec": "ts-node --project tsconfig.server.json server/index.ts",
+  "watch": ["server.ts"],
+  "exec": "ts-node --project tsconfig.server.json server.ts",
   "ext": "js ts"
 }
