--- tsconfig.server.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "dist",
-    "lib": ["es2019"],
-    "target": "es2019",
-    "isolatedModules": false,
-    "noEmit": false
-  },
-  "include": ["server/*.ts"]
-}
